[VDP] Protect palette setting from interrupt
[Code] Removed type casting to MSXgl version

+++ engine/src/core.h
@@ -8,9 +8,9 @@
 #pragma once
 
 // Library version
-#define VERSION_MAJOR				(u16)1		// 4-bits (0-15)
-#define VERSION_MINOR				(u16)4		// 6-bits (0-63)
-#define VERSION_PATCH				(u16)0		// 6-bits (0-63)
+#define VERSION_MAJOR				1		// 4-bits (0-15)
+#define VERSION_MINOR				4		// 6-bits (0-63)
+#define VERSION_PATCH				0		// 6-bits (0-63)
 
 // Macro: VERSION
 // Combines major, minor, and patch versions into a single version number.
